changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11738 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/CommandCenter/Scripting/ORNodeEvaluator.h
@@ -63,7 +63,7 @@
 - (int) functionLevel;
 - (NSString*) functionName;
 - (void) setValue:(id)aValue forIndex:(int) anIndex;
-- (unsigned) symbolTableCount;
+- (NSUInteger) symbolTableCount;
 - (id) symbolNameForIndex:(int)i;
 - (id) symbolValueForIndex:(int)i;
 - (NSMutableDictionary*) minSymbolTable;
